Add remove functionality

+++ episodic/src/lib/EpisodicList.js
@@ -2,35 +2,34 @@ const promisePool = require('../repositories/mysql');
 const User = require('./User.js');
 
 class EpisodicList {
-    constructor(name, id, obj) { 
-        if (obj)
-        {
+    constructor(name, id, obj) {
+        if (obj) {
             Object.assign(this, obj);
             return;
         }
 
         this.name = name;
-        this.podcasts = []; 
+        this.podcasts = [];
         this.episodes = [];
         this.id = id;
-      }
+    }
 
-      get listName () { return this.name; }
+    get listName() { return this.name; }
 
-      get listPodcasts () { return this.podcasts; }
+    get listPodcasts() { return this.podcasts; }
 
-      get listEpisodes () { return this.episodes; }
+    get listEpisodes() { return this.episodes; }
 
-      get listId () { return this.id; }
+    get listId() { return this.id; }
 
-      addPodcast(podcast) {
-          this.podcasts.push(podcast);
+    addPodcast(podcast) {
+        this.podcasts.push(podcast);
 
-          let podcastId = 1; //Get from database adapter probably
-          let sql = "insert into lists_podcasts_link (listsId, podcastsId) values (" + this.id + "," + podcastId + ")";
-          
-          //throw it in the database
-          new Promise(async (res, rej) => {
+        let podcastId = podcast.id;
+        let sql = "insert into lists_podcasts_link (listsId, podcastsId) values (" + this.id + "," + podcastId + ")";
+
+        //throw it in the database
+        new Promise(async (res, rej) => {
             try {
                 let result = await promisePool.query(sql);
                 res(result.insertId);
@@ -39,16 +38,23 @@ class EpisodicList {
                 res(err);
             }
         });
-      }
+    }
 
-      removePodcast(podcast) {
-          this.podcasts.pop(podcast);
+    removePodcast(podcast) {
+        let found = false;
+        this.podcasts.forEach((element) => {
+            if (element.id == podcast.id) {
+                this.podcasts.pop(element);
+                found = true;
+            }
+        });
 
-          let podcastId = 1; //Get from database adapter probably
-          let sql = "delete from lists_podcasts_link where listsId = " + this.id + " and podcastsId = " + podcastId + ";"
-          
-          //throw it in the database
-          new Promise(async (res, rej) => {
+        if (found == false) return;
+
+        let podcastId = podcast.id;
+        let sql = "delete from lists_podcasts_link where listsId = " + this.id + " and podcastsId = " + podcastId + ";"
+
+        new Promise(async (res, rej) => {
             try {
                 let result = await promisePool.query(sql);
                 res(result.insertId);
@@ -56,21 +62,46 @@ class EpisodicList {
                 console.log(err);
                 res(err);
             }
-          });
-      }
+        });
+    }
 
-      updatePodcast(podcast) {
-          this.removePodcast(podcast);
-          this.addPodcast(podcast);
-      }
+    updatePodcast(podcast) {
+        this.removePodcast(podcast);
+        this.addPodcast(podcast);
+    }
 
-      addEpisode(episode) {
-          this.episodes.push(episode);
+    addEpisode(episode) {
+        this.episodes.push(episode);
 
-          let episodeId = 1; //Get from database adapter probably
-          let sql = "insert into lists_episodes_link (listsId, episodesId) values (" + this.id + "," + episodeId + ")";
+        let episodeId = episode.id; 
+        let sql = "insert into lists_episodes_link (listsId, episodesId) values (" + this.id + "," + episodeId + ")";
 
-          new Promise(async (res, rej) => {
+        new Promise(async (res, rej) => {
+            try {
+                let result = await promisePool.query(sql);
+                res(result.insertId);
+            } catch (err) {
+                //console.log(err);
+                res(err);
+            }
+        });
+    }
+
+    removeEpisode(episode) {
+        let found = false;
+        this.episodes.forEach((element) => {
+            if (element.id == episode.id) {
+                this.episodes.pop(element);
+                found = true;
+            }
+        });
+
+        if (found == false) return;
+
+        let episodeId = episode.id; 
+        let sql = "delete from lists_episodes_link where listsId = " + this.id + " and episodesId = " + episodeId + ";"
+
+        new Promise(async (res, rej) => {
             try {
                 let result = await promisePool.query(sql);
                 res(result.insertId);
@@ -78,30 +109,13 @@ class EpisodicList {
                 console.log(err);
                 res(err);
             }
-          });
-      }
+        });
+    }
 
-      removeEpisode(episode) {
-          this.episodes.pop(episode);
-
-          let episodeId = 1; //Get from database adapter probably
-          let sql = "delete from lists_episodes_link where listsId = " + this.id + " and episodesId = " + episodeId + ";"
-
-          new Promise(async (res, rej) => {
-            try {
-                let result = await promisePool.query(sql);
-                res(result.insertId);
-            } catch (err) {
-                console.log(err);
-                res(err);
-            }
-          });
-      }
-
-      updateEpisode(episode) {
-          this.removeEpisode(episode);
-          this.addEpisode(episode);
-      }
+    updateEpisode(episode) {
+        this.removeEpisode(episode);
+        this.addEpisode(episode);
+    }
 }
 
 module.exports = EpisodicList;
